refactor: useDarkMode 훅 추출로 중복 코드 제거
