+++ app/about/_components/AboutRenderer.tsx
@@ -1,33 +1,15 @@
 'use client';
 
-import { useAtomValue } from 'jotai';
 import { type NotionBlock, Renderer } from 'notion-to-jsx';
-import { useEffect, useState } from 'react';
 
-import { themeAtom } from '@/atoms/theme';
+import useDarkMode from '@/hooks/useDarkMode';
 
 interface AboutRendererProps {
   blocks: NotionBlock[];
 }
 
 const AboutRenderer = ({ blocks }: AboutRendererProps) => {
-  const theme = useAtomValue(themeAtom);
-  const [isDarkMode, setIsDarkMode] = useState(false);
-
-  useEffect(() => {
-    if (theme === 'system') {
-      const systemDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
-      setIsDarkMode(systemDark);
-
-      const mediaQuery = window.matchMedia('(prefers-color-scheme: dark)');
-      const handleChange = (e: MediaQueryListEvent) => setIsDarkMode(e.matches);
-
-      mediaQuery.addEventListener('change', handleChange);
-      return () => mediaQuery.removeEventListener('change', handleChange);
-    }
-
-    setIsDarkMode(theme === 'dark');
-  }, [theme]);
+  const isDarkMode = useDarkMode();
 
   return (
     <Renderer
