+++ app/posts/[slug]/_components/PostRenderer.tsx
@@ -1,10 +1,8 @@
 'use client';
 
-import { useAtomValue } from 'jotai';
 import { type NotionBlock, Renderer } from 'notion-to-jsx';
-import { useEffect, useState } from 'react';
 
-import { themeAtom } from '@/atoms/theme';
+import useDarkMode from '@/hooks/useDarkMode';
 
 interface PostRendererProps {
   blocks: NotionBlock[];
@@ -13,23 +11,7 @@ interface PostRendererProps {
 }
 
 const PostRenderer = ({ blocks, title, cover }: PostRendererProps) => {
-  const theme = useAtomValue(themeAtom);
-  const [isDarkMode, setIsDarkMode] = useState(false);
-
-  useEffect(() => {
-    if (theme === 'system') {
-      const systemDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
-      setIsDarkMode(systemDark);
-
-      const mediaQuery = window.matchMedia('(prefers-color-scheme: dark)');
-      const handleChange = (e: MediaQueryListEvent) => setIsDarkMode(e.matches);
-
-      mediaQuery.addEventListener('change', handleChange);
-      return () => mediaQuery.removeEventListener('change', handleChange);
-    }
-
-    setIsDarkMode(theme === 'dark');
-  }, [theme]);
+  const isDarkMode = useDarkMode();
 
   return (
     <Renderer
